# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank management system
  description: api for managing user's cards
  contact:
    name: Panasik Uladzislau
    url: https://github.com/Vladislav3421730
    email: panasikvladislav1@gmail.com
  version: 1.0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /api/v1/limit/monthly:
    put:
      tags:
        - Limit Management
      summary: Set monthly limit
      description: Updates the user's monthly spending limit
      operationId: setMonthLimit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LimitRequestDto'
        required: true
      responses:
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/limit/daily:
    put:
      tags:
        - Limit Management
      summary: Set daily limit
      description: Updates the user's daily spending limit
      operationId: setDayLimit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LimitRequestDto'
        required: true
      responses:
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/transaction:
    get:
      tags:
        - Transaction Management
      summary: Get all transactions
      description: Retrieve a paginated list of all transactions
      operationId: findAll_2
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: minAmount
          in: query
          required: false
          schema:
            type: number
        - name: maxAmount
          in: query
          required: false
          schema:
            type: number
        - name: operation
          in: query
          required: false
          schema:
            type: string
        - name: operationResult
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully find transaction
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/transaction/{userId}/user:
    get:
      tags:
        - Transaction Management
      summary: Get transactions by user ID
      description: Retrieve transactions associated with a specific user ID
      operationId: findAllByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Successfully find transaction
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Transactions not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/transaction/{cardId}:
    get:
      tags:
        - Transaction Management
      summary: Get transactions by card ID
      description: Retrieve transactions associated with a specific card ID
      operationId: findAllByCardId
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Successfully find transaction
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Transactions not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/transaction/me:
    get:
      tags:
        - Transaction Management
      summary: Get current user's transactions
      description: Retrieve transactions of the currently authenticated user
      operationId: findAllUsersTransactions
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Successfully find transaction
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/transaction/me/{cardId}:
    get:
      tags:
        - Transaction Management
      summary: Get current user's transactions by card ID
      description: Retrieve current user's transactions for a specific card
      operationId: findAllUsersTransactionsByCardId
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Successfully find transaction
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Transactions not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/block/{userId}/users:
    get:
      tags:
        - Block Requests Management
      summary: Get block requests by user ID with pagination
      description: Returns a paginated list of block requests filtered by the specified user ID
      operationId: findAllByUserId_1
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Requests were found successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/block/{cardId}/cards:
    get:
      tags:
        - Block Requests Management
      summary: Get block requests by card ID with pagination
      description: Returns a paginated list of block requests filtered by the specified card ID
      operationId: findAllByCardId_1
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Requests were found successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
components:
  schemas:
    LimitRequestDto:
      type: object
      description: DTO for setting a transaction limit on a specific card
      properties:
        cardId:
          type: string
          format: uuid
          description: UUID of the card to apply the limit to
          example: 123e4567-e89b-12d3-a456-426614174000
        limit:
          type: number
          description: Limit value to be applied to the card
          example: 5000
          minimum: 0
      required:
        - cardId
        - limit
    FieldErrorDto:
      type: object
      description: DTO for errors with validation
      properties:
        code:
          type: integer
          format: int32
          description: Error's code
          example: 400
        timestamp:
          type: string
          description: Timestamp of when the error occurred
          example: '2025-02-07 14:30:00'
        errors:
          type: object
          additionalProperties:
            type: string
          description: Error map describing the validation
    AppErrorDto:
      type: object
      description: DTO for errors
      properties:
        code:
          type: integer
          format: int32
          description: Error's code
          example: 400
        timestamp:
          type: string
          description: Timestamp of when the error occurred
          example: '2025-02-07 14:30:00'
        message:
          type: string
          description: Error message describing the problem
          example: Invalid request
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModel:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          $ref: '#/components/schemas/PageMetadata'
...
  securitySchemes:
    BearerAuthentication:
      type: http
      scheme: bearer
      bearerFormat: JWT